SchemaParser: fetch schema tables sequentially
To minimize the impact of new client connecting to a cluster
with a large number of tables, this patch changes the way
schema tables are fetched by the driver.

The ideal solution would be to make the queries use paging but
that's not currently possible because of the driver architecture.

This is the first step to mitigate the problem. It makes the queries
run sequentially instead of all in parallel. Also schema data is no
longer fetched with single full scans. Instead we fetch keyspaces info
with a full scan and then fetch schema data for each keyspace separately
with a single partition reads. Keyspaces are fetched sequentially.

Signed-off-by: Piotr Jastrzebski <piotr@scylladb.com>